Increase the number of AI starting defenders, switch them to warriors, deprioritize archery early
Warriors are more versatile than archers, they can later be upgraded to axemen for example
and allow a more effective rush.
Only if the AI is militarily weaker and/or considering turtling should archers be more relevant.

Archery is also not that useful early to other techs that allow an early expansion.
I have switched it to pottery instead that should help the AI a lot more (early
granary or/and cottages).