give free techs to eligible players at all eras To make AIs keep up with the human player as the human player catches up in the later game, boosting the early game of the player getting free techs.
My approach is to give less valuable techs in terms of tech cost, but that
should critically help the AI early (either because of their high reward such
as currency enabling commerce for example sooner, or fishing and compass not
having time spent on them so the AI goes straight to the more effective techs
to have (while still allowing some variety).

The purpose would be to have a tenser and closer game

If this is too hard for the player(s) not receiving bonuses, there may be
other ways to help them while still helping the weaker player(s) staying
competitive for a closer game, even though they should still make decisive
moves to not lose against the other rivals.