Give explorer units a woodsman 1 free promotion
For the explorer:
- Unlike what is said in advc.315c, i think explore units should be good
at exploring. Navigating forest type terrains is a good way to do so.
- It may be more realistic and immersive this way.
- The verbose should not be a problem too.
- There should be better ways to rework the unit maybe than to remove some
of its charateristic abilities. I have not thought of such a way (yet?),
in all cases i see no strong reason to not allow this, so for now i am
reenabling woodsman 1 free promotion. I may or not later address this
point in advc.315c in another way, but not guaranteed and maybe it would
stay as is.

For the scout:
- Similarly to what i did for the explorer unit, i think explore units
should be good at exploring, navigating forest type terrains is a good
way to do so.
- It is also a way to make the woodsman promotion play more of a key role
in the game, as it is currently quite useless i think. Scouts being good
at such "terrains" (forest-jungle, that civ4 code seems to name features)
may be more realistic and immersive this way.

I did not think of it in the previous commit, but having a more balanced woodsman promotion
(and a bit stronger one too) may also help AIs who overinvest in it only and perhaps
do not use it the most efficiently (woodsman 3 aztec units never attacking in forest
or jungle, at least they would have some chance of being attacked while they navigate
the forest or the jungle), so all in all is maybe good even more.