reworked explorer units - much more specialized and perhaps more valuable now too
Reworked the explorer in a way that hopefully better addresses advc.315

Key points about the reworked explorer unit:
- much lower base strength: 2 (was 4)
- starts with woodsman 1 and 2, flanking 1 and 2
- can now attack all rivals and also cities (more realistic), but due
to their low strength unlikely they would make a difference. Woodsman
promotions also do not help for city attacks as city inner tile is currently
never forest or jungle, so it should be strongly disincentivized to do so
for human players, but maybe luckily they can capture rarely weakly guarded
cities of low strength units, it would be realistic, and risk losing the
high value explorer, so most likely not worth, but at least now not not
making sense that explorers could never attack other units, now it's just
not worth doing so with them.
- extra NotUnitAIs tags to make strictly and extra sure they are
not used as war units inadvertedly by the AI. As for the human player(s),
they should find a lot of value for using them as scouting units, but
- lower cost: 30 (was 40), it fits quite nicely at double the scout stats
- since AIs often have compass as free tech, makes this tech quite more
valuable

About the scout unit:
- slightly buffed it by giving the 1st tier promotions only unlike the
explorer, should add a few more minor changes in the next commit.