rework the quechua warrior, allow city garrison on melee Units - the quechua is a bit too weak in advciv i think, making it more versatile by basing it on the spearman now (which seems quite accurate historically (weapon type and long involvment in eras it seems) too), perhaps this is interesting maybe even as compared to the civ4 bts approach. It is now a cheap unit to build and very versatile, which may also be accurate from the little i read (that their power was mostly in numbers)
- gave now all melee units access to city garrison promotion.
No strong reason to not allow that, was initially for the quechua,
but melee units could defend cities, maybe is more realistic or/and
interesting this way, while adding more strategic depth now perhaps.
Will try to give defender units more of an edge at defending in other
ways maybe.