rework the holy roman special unit Similar to what was done to the reworked pikeman, but with a few changes to be more historically accurate, basing myself on what i've read, and also more balanced or/and challenging-interesting or/and immersive in the game maybe
Key differences include:
- Cost increased to 80: they are mercenaries not a nation's
dedicated unit, should be more expensive, plus they are stronger
than the regular reworked pikeman
- More terrain bonuses and penalties, see code comments for details.
They are stronger in terrains than the regular reworked pikeman,
to be more accurate and perhaps add strategic depth too.
- Available at civil service while still requiring feudalism (and
bronze working), this seems to be more historically accurate for
the era in which they appeared vs the pikemen, and also the fact
that they were recruited not serving a lord loyally so maybe more
about civil service. Also, since civil service is just after
feudalism, it is easier to get the holy roman units right after
feudalism and some other units too, so maybe this is more coherent
this way maybe.