fixed the math explanation of POWER_CORRECTION, added K-Mod's calculation's explanation too to double check should be much more robust now.
I found it weird that as i got closer to K-Mod's value for
POWER_CORRECTION of 166 (increasing from 159 to 160), instead
of getting closer to the K-Mod's value of 3.16 warriors per
spearman (i had 3.15 due to not rounding, see below, now
"fixed"or rather changed as it was not necessarily an issue
i mean anyways, but indeed slightly less correct at least in
this case i mean anyways), it (from 3.010 warriors per spearmen
for a POWER_CORRECTION of 159, to 2.990 (wrong but did not
notice at that time in this case and time at least i mean anyways)
warriors per spearmen for a POWER_CORRECTION of 160) drove me further
away from K-Mod's value 3.016 warriors per spearmen for a
POWER_CORRECTION of 166, which did not align intuitively and was
weird if i may say indeed in this case too maybe i mean anyways.

So now I fixed this by adding rounding so i have 3.16 (= 3.160 too)
and not 3.15 (i had found 3.15 initially but it was due to not rounding
4^1.66 = 9.9866 to 9.987 (i had kept instead 9.986) now i find the
same value as in K-Mod).

And from there i also applied rounding to all my values, which may have
slightly changed the calculation but not a lot i mean anyways at least
in this case i mean.
While doing so, rechecking all my values, and now the warriors per
spearmen values, when going from a POWER_CORRECTION of
159 to 160 (vs 166), are now leaning towards K-Mod's value's direction
at least in this case i mean anyways:
3.011 to 3.032 (vs 3.160)

So now the math should be very robust unless i made a huge mistake
or something in this case i mean anyways.
I also think the explanation helps in the code, even though is a bit
verbose or too long, if i may say i mean anyways, or maybe not, at
least in this case i mean anyways maybe, as it can be also helpful
if some people want to modify this value i mean anyways.