revert most CIV4PromotionInfos.xml changes, keep the ones i find most valuable In this case i mean anyways, also there are quite a bit of changes i kept in this case i mean anyways.
I think many are worth merging in advciv, but awaiting review for
that to have your opinion.
If you were not to be interested in these, so be it, it would be as
it is, also about this case maybe i mean anyways.
I think they are quite worthwhile about this i mean anyways, rest
depends on the review, i mean anyways, else is as is i mean anyways
at least in this case, thanks,