reduce self.W_CATEGORIES from 230 to 200 (in SevoPediaMain.py) is maybe a bit extreme but i badly need/want the place, especially for sevopedialeader, the diff is quite nice, more info is displayed now, screenshots appended are from 230 to 210 (more/also are available in the google drive, anyways), further gains going to 200.
199 Does not work/function correctly as "Concepts (Outdated)" now
takes 2 lines/row to display, and i really need the "Outdated" word
or want to display it, is maybe a fine point to stop at
anyway/=regardless as other categories need space/word too.

I said is extreme (a bit i said too the word but anyways) but maybe
is not, these don't serve much purpsoe except displaying the category
header labels which they do, i'd rather take the space in the sevopedia
that is directly useful adn helpful hopefully but anyways.

May increase it a bit to 205 or 210 for example for visual clarity or
smoothness, but not guaranteed and may not, this is quite fine overall,
anyways, and i can use the place for other things too or/and not other
things or/and not or/and yes or/and not etc, anyways,