partially add/display the changing among all leaders few (4 to be exact anyways etc) ACLs ideally would have lvoed to do a deeper affeciton resentment as said in more detial in the code comments (added (too (anwyays etc)))  in ai_attributes_displayed_config.py (please/you can/may(?) see them/these for details anyways etc) in this commit
But it does also display some useful information, and it is also true the other attitude
changes +/- limits +/- divisors don't change among leaders, not ideal, maybe to change or
not, but in all cases hoepfulyl helpful or maybe not but in all cases or not anyways etc;
it does also wrap up the sevopedia leader panel quite nicely, while adding more relevant
info and sorting it a bit too, anyways etc.

Minor other changes or/and fixes in this commit anywyas etc.