revert to using default civ4 art assets for daoism, keeping the renamed xml tags though no need to create further issues as we use default assets anyways, but it was working fine for a few and most art assets actually (only temple, monastery, and (daoist (too anyways etc) missionary had an issue (purple animation(/"sprite"?) but otherwise displayed fine, still as we don't need them revert to default and workign files (we don't modify them))))
as for the added files, we don't have a strong reason to remove them, and
they are necessary enough (i wanted to say this thought i had a bit earlier
is maybe bit quite cool or not yes or other hehe but anwyays etc is as and
what i wanted to say for better or worse or other or not but anyways etc
anyways etc)

as for the art assets themselves, i had imported them using dragon unpacker,
which seemed to had worked fine:
2025/05/07 17:20:01.754 [M] Opening "C:\Program Files (x86)\Steam\steamapps\common\Sid Meier's Civilization IV Beyond the Sword\Assets\Art0.FPK" file:
2025/05/07 17:20:01.758 [M] Driver plugin "Elbereth's Main Driver" thinks it can open this file.
2025/05/07 17:20:01.760 [H] Opening the file using "Elbereth's Main Driver" plugin... 93ms Success!
2025/05/07 17:20:01.863 [H] Parsing entries for directories... 120ms
etc...

I had also noticed after it that an entry (taoism.dds) (renamed to daoism.dds
before reverting this change now in this commit) was duplicated and needed to
be exported separately, but the purple "sprites"(?) issue could not be easily
resolved (by me but anyways etc) is fine as it is, by that i mean the game
works/fucntions and such with default .dds and such all daoism assets with
still the old name, probably for the best or not yes or and other or not,
anyways etc.

At least they are accessible if want in this commit, for example if i (as
expected here anyways etc) add more religions, paganism being next and replacing
old confucianism, so maybe all useful and nice or not, or and other or and not or
yes or and other or and not, anyways etc anyways etc, now works/fucntions fine,
anyways etc.