major barbarian rework, + other changes major barbarian reword i am very happy about but anyways etc, tested it (see the drive for preview images anyways etc but anyways etc) and it seems stable and quite balanced.
The changes are overall that:
- Barbarians should now be much stronger and relevant in the early game,
they should tech and pace at about same pace if not slightly faster than
average in the early game, and eventually around turn 150 to 250 they
start to finally collapse and be conquered by other civs as they can't
keep up, which i find both realistic and nice anyways etc
- This is achieved by treating barbarian leader still not as a real player
(no diplomacy and such if there is such or not anyways etc), but they get
a few custom((/)i.e. civ-specific) buildings that cost 1 hammer and with
sometimes different stats but rarely if not never or not or and other or
and not anyways etc besides that but anyways etc ; including in particular
a free palace in every city that is given for free at each other barbarian
building being built (redundant but maybe not incorrect (but) anyways
etc...(!!(!) anyways etc...))
- This is balanced and not op, as they have a later start, sometimes at
turn 100+ for some cities, and while they also get access to all military
units now (and a few custom units (worker, workboat) at a 1 hammer cost
version of these too, just to speed up their growth or rather not slow it
down so they catch up their delay anyways etc), and some buildings too if
they can and it is relevant i found for them to build (i want them to keep
a warmonger profile, but not expressely hardly trying to prevent their
growth in a quite general sense otherwise, only if i find it less efficient
or not worth it for them or and me or not or yes or and other or and not
anyways etc anyways etc), they still stagnate around city size 8-10 (could
be adjusted by increasing the "smileys" (anyways etc...) in their barbarian
palace(s since all cities have a palace now, which as an unnoticed side
effect that i didn't intend at first but noticed later but anyways etc
addresses their cities being so sparse quite nicely if i am not mistaken
anyways etc))
- Handicap has also been adjusted so there are less combat bonuses against
them, for a more fair and even game, and fun perhaps too but just my opinion
but anyways etc, so my goal (or/and feel/wish but anyways etc) is to make
them stronger but not op, scale better and be relevant longer in the game,
also having a few more cities now, which also (happens to but anyways etc)
settle the unsettled/unplanted areas/zones of the map quite nicely anyways
etc, so even at lower difficulties they should be (quite a) challenge (hehe
(but)(anyways etc...)) but hopefully not too hard but not too easy either,
i did my tests at emperor and did not finish balancing difficulties not
actually started too indepth yet but that i would want to do hopefully later
but anyways etc. So far these goals seem to be achieved quite well and i am
very satisfied of the result, please look at notes for details or code or/and
comments.

Hopefully it is more fun to play this way, but anyways etc.

Other changes:
- renamed OBELISK to MONUMENT when relevant or to EGYPTIAN_OBELISK
when relevant, and vice versa if i am not mistaken too renamed
MONUMENT or not but maybe rather look at code comments where i
explained it very well (as in in depth if not too spammy maybe
hopefully or not or verbose or not hehe but i like this way but
anyways hopefully easy enough to read or not or yes or not or
and other or and not anyways etc)
- updated docs for these barbarian changes too anyways etc
- renamed CIVIC_PAGANISM to CIVIC_PREHISTORIC_RELIGION to not overlap
with our now paganism religion, please see code for details if need
be and the wiki or maybe rather sevopedia anyways etc pedia entry
modified from wiki or and other places or maybe not here in this
case or not in this case or not or yes or not or and other or and
not anyways etc (but) anyways etc...
- added an empty TXT_KEY_DUMMY_EMPTY_CUSTOM_TXTKEY_STRATEGY txt key
but anyways etc to not clutter with uneeded (and) unused (but)
anyways etc, not sure it is necessary as we don't use the fields at
all, but maybe we respect the xml architecture this way (of civ4 i
mean anyways etc), i did not check if we applied it eveyrwhere needed
if it is even needed but etc anyways etc anyways etc, but anyways etc
- minor other changes or/ad other i may have forgotten or/and not,
hopefully accurate enough, please view commit diff for details,
sometimes it seems the github page needs to be refreshed to show them,
sometimes a few times else it trims with connection error or something
some (other?) kind of error bu anyways, not guaranteed and i could be
mistaken or maybe not, but anyways etc