add placeExclusiveCivs civilization(s) button(s) in sevopediaunit thanks to Claude AI (is instead of textual link too anyways etc), add info about a bug of new game screen being stuck about fastdep compile and seemingly solved after erasing/deleting temp_files and recompiling, reduce ibuildunitprob to 70 for barbarians to try to allow them to build more non-war things anyways etc updated docs about Claude AI too anyways etc
about the weird screen being stuck bug, not sure but should be maybe but
not guaranteed, see docs for details (there are google drive links in
particular with screenshot and the suspected DLL, as well as a longer
explanation of the issue and how it seems to now have been fixed and
why see there for details i mean if i may say etc but anyways etc
anyways etc)

Reduced ibuildunitprob (of barbarian leader anyways etc) to a lower value
(here 70 for example anyways etc) to allow them or attempt to allow them
(barbarians anyways etc) to build a few nice things if i may say i mean
anyways or maybe not nice or nice or and other or and not anyways etc,(;?
Anyways etc... anyways etc) such as work boats if city is naval, barbarian
lighthouse, perhaps a few barbarian workers too if it helps, not sure
reducing this value will for sure allow these (as for example barbarian
workers seem to alwys stay in city almost (maybe due to fear of capture
due to war with all civs if i am not mistaken? That i would want to
ideally change but that functionally is not too bad either maybe as
players cant steal barbarian workers maybe too anyways etc)), but in
all cases trying this value and see(ing) anyways etc anyways etc.

or/and other minor changes or/and doc updates or/and other or and not
anyways etc