renamed BONUS_WINE to BONUS_GRAPES, rework grapes, now based on IMPROVEMENT_FARM, removed IMPROVEMENT_WINERY, reworked BUILDING_GROCER('s) association with grapes and sugar (now gives happiness not health), update grapes button (Realism Invictus mod import) As said quite nicely if i may say but anyways etc and requoting myself anyways etc anyways etc anyways etc... From ((/I?)in(Anyways etc)) the sevopedia changes entry or maybe rather the README_Quick_Get_Started_Guide.md that should be quite similar if not very very but without the sevopedia links in it anyways etc:
"
- Renamed "Wine" (Bonus) to "Grapes". As part of renaming wine to grapes, additionally reworked grapes: now improved by IMPROVEMENT_FARM, not anymore by the old improvement_winery now removed. This matches historicality better if i may say and if it's a word anyways etc, as wine was consumed much before 4000 BC. Also, some wines are not made from grapes at all if i am not mistaken, and grapes can be eaten out of the wine usage. Updated bonus's description with a wikipedia based one as often if not always done when updating pedia entries of/in advciv-sas or and other mods using this mod's (i.e. this advciv-sas's code anyways etc). Also, the winery/happiness part of wine is transferred to the grocer building instead, now grapes gives happiness not health anymore. Similarly although not directly related, sugar now also gives happiness not health anymore through grocer. Hopefully buffes these mid game bonus effects while making these ressources more relevant as they were (a) bit (anyways etc) weak compared to say wheat or corn. See code comments for details or the ingame sevopedia entry of the grapes with also maybe or not or yes anyways etc latest updated info if anything changed since i wrote (am writing currently but anyways etc also wrote when viewing from later the new now maybe or not or yes anyways etc) but this should be the bulk of it and hopefully makes gameplay more accurate or/and interesting perhaps or/and enjoyable or not anyways etc.
"

Note: about the commit message('s content) that i said before (but anyways etc anyways etc anyways etc) in last/previous commit, i want to be extra clear and careful but anyways etc anyways etc anyways etc, that i strictly meant it only about civ4 content, however, as also my mind can't help but to wander (and wonder even though unrelated or not but anyways etc) and overthink (which i like too hehe but anyways etc anyways etc anyways etc), i may have written this with other thoughts in mind, but in the end i only intended it to be about civ4 content, at least at that time if not always or not or yes or other or not but maybe yes only civ4 content indeed anyways etc. But it cannot strictly be said either that i don't think about unrelated or not directly related "parallel" thoughts, maybe is how my mind is which i like too but anyways etc anyways etc anyways etc, as for this message this was to be extra clear about what it meant and what i meant by it. But it is also something i felt (intuitively i think? As my mind is too which i like too indeed, wasn't it (also intuitive)? But anyways etc anyways etc anyways etc...), so i just did so, not to minimize its importance, just i didn't want to hold back whatever comes of it, but in the end i only speak for myself, and with that said, and the clarification i wanted to make made hopefully, with its nuance or not nuance, but anyways etc, and quite clearly that i only meant it about civ4 content and nothing else, hopefully clear enough to myself too perhaps anyways etc, now continuing civ4 modding or/and othe rhtings i want to do or and do or and not as i prefer or not or do or not or and other or and not anyways etc. Ths is also to not (over)stress myself, but it didn't accomplish so exactly, still i wanted to clarify this, maybe it helped a bit or not is for me to manage for the part of the clarification's feeling that is on my end, in all cases continuing civ4 moddingor not or and other or and not as i prefer or nto or do or not, not as an excuse or cover, but as what i want to do, not to sound too generic or "barriered"/procedural or soemthign which i don't like to be whatever i am or would be is what i am or would be, but or rather maybe with that all said as i wanted, hopefulyl clearly enough r not, anyways etc, continuing (i (or those who like to follow my changes and such whch i like to share too hehe but anyways etc, nt sharing always all or not but wnays etc (as is tedious bit sometimes hehe, and sometimes i don't want to share all too even if not tedious... but anyways etc... Trying to shar emost when i want it or do or not, not promised or guaranteed, just what i would maybe or not maybe often or not often do or not do, anyways etc hopefully clear enough or clearer hopefully maybe rather or and helpful or not but in all cases with that said anyways etc...)))