revert the nicolor static const caching optimization i had added i had added it in https://github.com/wonderingabout/AdvCiv-SAS/commit/00866a0741a1e0818a9d9696d120cd9cd4551839#diff-3b9a656021f740cb1ac975c5b4cfc3d7340fc6b8c7633c5e8b6a18aa633013b2R2234-R2247
unlike the static const colortypes caching, i am not sure these are safe, and in general i don't know too much if at all about these.

Chatgpt i don't know which version (but maybe/most likely 5 since it was released at that time if i'm not mistaken but anyways etc) said it was safe, now chatgpt 5 tells me better revert it.

We had a weird crash when clicking on main menu button ingame in map view, so maybe this is related. Even if not, the advantage is not worth the bother anyways etc.