removed `UNITAI_CITY_COUNTER` from non-versatile enough units such as the UNIT_MACEMAN, as they are not versatile enough to be reliable city defenders while still being good offense units - removed `UNITAI_CITY_COUNTER` from non-versatile enough units such as the UNIT_MACEMAN, as they are not versatile enough to be reliable city defenders while still being good offense units (too easy to punish with a stack of knight units that are good against them, even a crossbow stack could defeat defending macemen, so use them for offense rather and disable them from here as well as in similar places anyways etc). They may be good against pikemen, but pikemen don't often attack cities and are not that efficient for that, so prefer macemen to focus more on offense unitai roles if i am not mistaken but anyways etc, and same for other units if any have that issue as well anyways etc.
Vs previous code, we also undid a lot of not relevant UNITAI_CITY_COUNTER assignments that caused for example excessive longbowmen production both with UNITAI_CITY_DEFENSE and UNITAI_CITY_COUNTER, which is very inefficient and suboptimal early. Allow longbowmen to be versatile if AI chooses to do so i mean, but otherwise try to not overproduce them as there are better/stronger offense units generally if we have defended our cities well enough if i am not mistaken, should be more efficient as such if i am not mistaken but anyways etc.

The macemen change is from how it was in base advciv though so i focused on it in this commit, but i also undid various advciv-sas changes to unitai related to UNITAI_CITY_COUNTER in xml, hopefully stronger and more relevant/efficient as such while still being versatile enough, which seems to be the case in autoplay (less excess longbowmen, varied composition otherwise if i'm not mistaken but anyways etc).