minor fix (iAIAnimalBonus at handicap_chieftain anyways etc) and update docs note: the released AdvCiv-SAS 5030 does not use the updated handicap setting (iAIAnimalBonus at handicap_chieftain anyways etc) in the development version as of now 5031, but the difference is so small i am not / will not update the released version 5030 just for that. I hope the documentation is accurate enough, and it always reflects latest development version, which as of now is exact same as released AdvCiv-SAS 5030 minus this difference if i'm not mistaken but anyways etc.
I prefer to give players updated doc even if it references the development version. Maybe i'll change it in the future in this case i mean but anyways etc but as of now documentation links to latest development version, not latest stable, although i could do so. I hope it is not too inconvenient or inaccurate as such, as it is much more convenient for me to do so, perhaps preferable or not to players as they prefer (redundant but anyways etc), but in short this is a very very minor fix, use dev version if you prefer, as for doc i'll keep referencing dev version as it is more accurate, especially if things changed, i prefer players to keep in mind the latest version of information as it will eventually (in theory f i eventually release/update it but anyways etc) become the next version anyway, in this case i mean but anyways etc.

Also updated docs anyways etc.