patch/fix tech beelines at tech_physics (to tech_steam_power), and tech_theory_of_evolution (to tech_genetics and tech_game_theory) and tweak tech tree as well while doing so i mean but anyways etc - add tech_naturalism as a prereq to tech_physics: "<!-- custom: nicely fits copernic theme and fighting previous religious or such preconceptions with the power of reality i mean if i may say for better or worse xd i mean but anyways etc -->" - add tech_electricity as a prereq to tech_genetics: "<!-- custom: added this extra as of now tech_electricity prereq nicely i mean if i may say but anyways etc recommended by chatgpt 5 but i also mean i mean that i got same idea independently while asking it and it was replying i mean if i may say but anyways etc, thanks for help and confirming lot! I mean if i may say but anyways etc. Also, as i explained to chatgpt 5 as it asked me i mean if i may say but anyways etc, it (i.e. tech_nationalism i mean but anyways etc) also seems to be a great fit i mean but anyways etc as this is a 1900s tech and actually a bit more, plus it is military themed and gives as of now permanent alliances in particular so we may want a far enough gate to prevent researching it (tech_theory of evolution is still available early for versatility or strategic variation (also preserves the tech_theory_of_evolution gambit (i.e. going for a tech_theory_of_evolution rush for later gains, skipping military path for some time) i mean as well but anyways etc) though i mean if i may say but anyways etc), and tech_liberalism is a bit too soon for that (tech chatgpt 5 sugegsted as an or to tech_nationalism) unlike tech_nationalism which fits thematically and in year time i mean if i may say but anyways etc. -->" - add tech_nationalism as a prereq to tech_game_theory: "<!-- custom: got the idea while i was asking chatgpt 5, to prevent a beeline that would be bit op or suboptimal depending on tech path maybe i mean but anyways etc, here i mean. Hopefully more balanced as such, and is also thematically relevant i think. Trying not to restrict strategic options (in particular this also preserves the tech_theory_of_evolution gambit i mean as well but anyways etc)  too much here for variety or such i mean as well but anyways etc -->"
all in all i mean if i may say but anyways etc, this nicely fixes the beelines while keeping strategical versatility hopefully i mean too but anyways etc. I tried more aggressive changes but it seems the tech tree is worse with them, so i hope the tech tree is balanced as such before a new release, maybe perfect even, although could always be improved maybe but the "when there is nothing left to take away" seemed to make sense here i guess i mean and felt.

Hopefully i can (not promised nor guaranteed, i may or i may not but anyways etc) ship a few fixes to AI or gameplay or such and make a new release after i have tested this enough (or at least a bit) (again not guaranteed nor promised, i may or may not do so and it may be long or short until then in this case i mean and if i ever do i mean but anyways etc).