slightly reduce costs in the early-mid game, and significantly more in the middle part of the game + fix a beeline at TECH_MACHINERY anyways etc. 325->300 450->425 600->575 800->750 1100->900 1400->1200 1800->1500 1900->1800 2000->1950
as recommended by gemini pro 3 and with my own ideas i mean if i may say but anyways etc.

Previous settings were a bit too slow, especially the transition from mid classical to renaissance.

But overdoing it makes ancient era too fast which is not pleasant either. So after some testing i have targeted most cost "curve" correction towards the middle game, while slightly increasing costs before that in some parts of the early game. Hopefully preserves a long enough early, but bit faster at the key mid classical parts but anyways etc, while still preserving a long enough ancient else it's just too fast; and then targetting more heavily the middle ages and slightly before which was slowest and which is where we changed the tech tree the most since AdvCiv-5082 if i'm not mistaken but anyways etc. (in advciv-sas 5082 we had 2 less columns (in medieval and the other one in renaissance if i'm not mistaken anyways etc.) so tech costs were cheaper than in the current development version anyway, this makes it closer to AdvCiv-SAS 5082 with this change while also making the early slightly faster as well as it was a bit too slow if i'm not mistaken i mean but anyways etc).

- TECH_MACHINERY: add TECH_IRON_WORKING as a prereq to prevent a beeline, plus also makes sense thematically to use iron for machines if i'm not mistaken but anyways etc.